11-Final-Project/docker/webapp/requirements.txt
streamlit
pandas
plotly
sqlalchemy
psycopg2-binary
matplotlib
